// src/generated/models/Mserp_ecoresproductcategoryentitiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Mserp_ecoresproductcategoryentitiesmserp_iscategoryinheritingparentcategoryattributes = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_ecoresproductcategoryentitiesmserp_iscategoryinheritingparentcategoryattributes = keyof typeof Mserp_ecoresproductcategoryentitiesmserp_iscategoryinheritingparentcategoryattributes;
export const Mserp_ecoresproductcategoryentitiesmserp_iscategoryinheritingparentproductattributes = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_ecoresproductcategoryentitiesmserp_iscategoryinheritingparentproductattributes = keyof typeof Mserp_ecoresproductcategoryentitiesmserp_iscategoryinheritingparentproductattributes;
export const Mserp_ecoresproductcategoryentitiesmserp_istangibleproduct = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_ecoresproductcategoryentitiesmserp_istangibleproduct = keyof typeof Mserp_ecoresproductcategoryentitiesmserp_istangibleproduct;

export interface Mserp_ecoresproductcategoryentitiesBase {
  mserp_categorycode?: string;
  mserp_categorydescription?: string;
  mserp_categorykeywords?: string;
  mserp_categoryname: string;
  mserp_defaultprojectglobalcategoryid?: string;
  mserp_ecoresproductcategoryentityid: string;
  mserp_externalid?: string;
  "mserp_FK_ParentProductCategory_id@odata.bind"?: string;
  mserp_friendlycategoryname?: string;
  mserp_iscategoryinheritingparentcategoryattributes?: Mserp_ecoresproductcategoryentitiesmserp_iscategoryinheritingparentcategoryattributes;
  mserp_iscategoryinheritingparentproductattributes?: Mserp_ecoresproductcategoryentitiesmserp_iscategoryinheritingparentproductattributes;
  mserp_istangibleproduct?: Mserp_ecoresproductcategoryentitiesmserp_istangibleproduct;
  mserp_parentproductcategorycode?: string;
  mserp_parentproductcategoryname?: string;
  mserp_pkwiucode?: string;
  mserp_primaryfield?: string;
  mserp_productcategoryhierarchyname: string;
  mserp_projectcategoryname?: string;
}

export interface Mserp_ecoresproductcategoryentities extends Mserp_ecoresproductcategoryentitiesBase {
  mserp_categoryrecordid?: number;
  mserp_categoryrecordid_bigint?: number;
  mserp_fk_parentproductcategory_idname?: string;
  mserp_iscategoryinheritingparentcategoryattributesname?: string;
  mserp_iscategoryinheritingparentproductattributesname?: string;
  mserp_istangibleproductname?: string;
  mserp_parentproductcategoryhierarchyname?: string;
  mserp_fk_parentproductcategory_id?: object;
  _mserp_fk_parentproductcategory_id_value?: string;
}