// src/generated/models/Mserp_ecoresproductcategoryassignmententitiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export interface Mserp_ecoresproductcategoryassignmententitiesBase {
  mserp_displayorder?: number;
  mserp_ecoresproductcategoryassignmententityid: string;
  "mserp_FK_ProductCategory_id@odata.bind"?: string;
  "mserp_FK_ProductV2_id@odata.bind"?: string;
  mserp_primaryfield?: string;
  mserp_productcategoryhierarchyname: string;
  mserp_productcategoryname: string;
  mserp_productnumber: string;
}

export interface Mserp_ecoresproductcategoryassignmententities extends Mserp_ecoresproductcategoryassignmententitiesBase {
  mserp_fk_productcategory_idname?: string;
  mserp_fk_productv2_idname?: string;
  mserp_fk_productcategory_id?: object;
  _mserp_fk_productcategory_id_value?: string;
  mserp_fk_productv2_id?: object;
  _mserp_fk_productv2_id_value?: string;
}