// src/generated/models/Mserp_ecoresproductcategoryhierarchyroleentitiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Mserp_ecoresproductcategoryhierarchyroleentitiesmserp_hierarchyrole = {
  200000000: 'Procurementcategoryhierarchy',
  200000001: 'Salescategoryhierarchy',
  200000002: 'Commerceproducthierarchy',
  200000003: 'Commoditycodehierarchy',
  200000004: 'Derivedfinancialhierarchy',
  200000005: 'Supplementalretailhierarchy',
  200000006: 'Retailvendorproducthierarchy',
  200000007: 'Retailchannelnavigationhierarchy',
  200000008: 'Packingmaterialhierarchy',
  200000009: 'Commondataservice',
  200000010: 'Costmanagementhierarchy',
  200000011: 'Traceabilityhierarchy',
  200000012: 'EngineeringProductHierarchy'
} as const;
export type Mserp_ecoresproductcategoryhierarchyroleentitiesmserp_hierarchyrole = keyof typeof Mserp_ecoresproductcategoryhierarchyroleentitiesmserp_hierarchyrole;

export interface Mserp_ecoresproductcategoryhierarchyroleentities {
  mserp_ecoresproductcategoryhierarchyroleentityid: string;
  mserp_hierarchyrole: Mserp_ecoresproductcategoryhierarchyroleentitiesmserp_hierarchyrole;
  mserp_hierarchyrolename?: string;
  mserp_primaryfield?: string;
  mserp_productcategoryhierarchyname: string;
}