// src/generated/models/Mserp_ecoresreleasedproductdocumentattachmententitiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Mserp_ecoresreleasedproductdocumentattachmententitiesmserp_accessrestriction = {
  200000000: 'Internal',
  200000001: 'External'
} as const;
export type Mserp_ecoresreleasedproductdocumentattachmententitiesmserp_accessrestriction = keyof typeof Mserp_ecoresreleasedproductdocumentattachmententitiesmserp_accessrestriction;
export const Mserp_ecoresreleasedproductdocumentattachmententitiesmserp_isdefaultproductimage = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_ecoresreleasedproductdocumentattachmententitiesmserp_isdefaultproductimage = keyof typeof Mserp_ecoresreleasedproductdocumentattachmententitiesmserp_isdefaultproductimage;
export const Mserp_ecoresreleasedproductdocumentattachmententitiesmserp_isproductimage = {
  200000000: 'No',
  200000001: 'Yes'
} as const;
export type Mserp_ecoresreleasedproductdocumentattachmententitiesmserp_isproductimage = keyof typeof Mserp_ecoresreleasedproductdocumentattachmententitiesmserp_isproductimage;
export const Mserp_ecoresreleasedproductdocumentattachmententitiesmserp_productimageusage = {
  200000000: 'External',
  200000001: 'Internal'
} as const;
export type Mserp_ecoresreleasedproductdocumentattachmententitiesmserp_productimageusage = keyof typeof Mserp_ecoresreleasedproductdocumentattachmententitiesmserp_productimageusage;

export interface Mserp_ecoresreleasedproductdocumentattachmententitiesBase {
  mserp_accessrestriction?: Mserp_ecoresreleasedproductdocumentattachmententitiesmserp_accessrestriction;
  mserp_attacheddatetime?: string;
  mserp_attachinguserid?: string;
  mserp_attachment?: string;
  mserp_attachmentdescription: string;
  mserp_dataareaid?: string;
  "mserp_dataAreaId_id@odata.bind"?: string;
  mserp_documentattachmenttypecode: string;
  mserp_documentattachmenttypelegalentityid: string;
  mserp_ecoresreleasedproductdocumentattachmententityid: string;
  mserp_filetype?: string;
  "mserp_FK_DocumentAttachmentType_id@odata.bind"?: string;
  "mserp_FK_ReleasedProductV2_id@odata.bind"?: string;
  mserp_isdefaultproductimage?: Mserp_ecoresreleasedproductdocumentattachmententitiesmserp_isdefaultproductimage;
  mserp_isproductimage?: Mserp_ecoresreleasedproductdocumentattachmententitiesmserp_isproductimage;
  mserp_itemnumber: string;
  mserp_notes?: string;
  mserp_originalfilename?: string;
  mserp_primaryfield?: string;
  mserp_productimageusage?: Mserp_ecoresreleasedproductdocumentattachmententitiesmserp_productimageusage;
}

export interface Mserp_ecoresreleasedproductdocumentattachmententities extends Mserp_ecoresreleasedproductdocumentattachmententitiesBase {
  mserp_accessrestrictionname?: string;
  mserp_dataareaid_idname?: string;
  mserp_filelocation?: string;
  mserp_fk_documentattachmenttype_idname?: string;
  mserp_fk_releasedproductv2_idname?: string;
  mserp_isdefaultproductimagename?: string;
  mserp_isproductimagename?: string;
  mserp_productimageusagename?: string;
  mserp_dataareaid_id?: object;
  _mserp_dataareaid_id_value?: string;
  mserp_fk_documentattachmenttype_id?: object;
  _mserp_fk_documentattachmenttype_id_value?: string;
  mserp_fk_releasedproductv2_id?: object;
  _mserp_fk_releasedproductv2_id_value?: string;
}