// src/generated/models/Mserp_ecoresproductmastercolorentitiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export interface Mserp_ecoresproductmastercolorentities {
  mserp_displaysequencenumber?: number;
  mserp_ecoresproductmastercolorentityid: string;
  mserp_primaryfield?: string;
  mserp_productcolorid: string;
  mserp_productmasternumber: string;
  mserp_replenishmentweight?: number;
  versionnumber?: number;
}