// src/generated/models/Mserp_ecoresproductmastersizeentitiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export interface Mserp_ecoresproductmastersizeentities {
  mserp_displaysequencenumber?: number;
  mserp_ecoresproductmastersizeentityid: string;
  mserp_primaryfield?: string;
  mserp_productmasternumber: string;
  mserp_productsizeid: string;
  mserp_replenishmentweight?: number;
  versionnumber?: number;
}