// src/generated/models/Mserp_ecoresproductmasterconfigurationentitiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export interface Mserp_ecoresproductmasterconfigurationentities {
  mserp_containerunitsymbol?: string;
  mserp_displaysequencenumber?: number;
  mserp_ecoresproductmasterconfigurationentityid: string;
  mserp_primaryfield?: string;
  mserp_productconfigurationid: string;
  mserp_productmasternumber: string;
  mserp_replenishmentweight?: number;
  versionnumber?: number;
}