// src/generated/models/Mserp_ecoresproductmasterstyleentitiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export interface Mserp_ecoresproductmasterstyleentities {
  mserp_displaysequencenumber?: number;
  mserp_ecoresproductmasterstyleentityid: string;
  mserp_primaryfield?: string;
  mserp_productmasternumber: string;
  mserp_productstyleid: string;
  mserp_replenishmentweight?: number;
  versionnumber?: number;
}